requirements.txt
asgiref==3.3.4
Django==3.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-crispy-forms==1.11.2
django-credit-cards==0.4.1
Pillow==8.2.0
